issue#16:申し込み完了メール本文の修正
