+++ package.json
@@ -18,7 +18,6 @@
     "ect": "^0.5.9",
     "ejs": "^2.5.6",
     "express": "^4.15.0",
-    "express-session": "^1.15.1",
     "express-validator": "^3.1.2",
     "forever": "^0.15.3",
     "moment": "^2.17.1",
