+++ public/javascripts/service.js
@@ -84,7 +84,7 @@ angular.module('myApp')
                         tel: $scope.tel,
                         email: $scope.email,
                         date: $scope.date,
-                        hour: $scope.hour,
+                        time: $scope.time,
                         attendNum: $scope.attendNum
                     }
                 }).
